requirements.txt
Appium-Python-Client==4.1.0
assertpy
requests==2.32.3
selenium==4.24.0